Moved Direction, Rotation, and MovecraftLocation to movecraft-api.
