+++ movecraft-api/src/main/java/net.countercraft.movecraft.api/MovecraftLocation.java
@@ -15,7 +15,7 @@
  *     along with Movecraft.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.countercraft.movecraft.utils;
+package net.countercraft.movecraft.api;
 
 public final class MovecraftLocation {
     public final int x, y, z;
