+++ movecraft-core/src/main/java/net/countercraft/movecraft/async/translation/TranslationTaskData.java
@@ -20,7 +20,7 @@ package net.countercraft.movecraft.async.translation;
 import net.countercraft.movecraft.utils.EntityUpdateCommand;
 import net.countercraft.movecraft.utils.ItemDropUpdateCommand;
 import net.countercraft.movecraft.utils.MapUpdateCommand;
-import net.countercraft.movecraft.utils.MovecraftLocation;
+import net.countercraft.movecraft.api.MovecraftLocation;
 
 public class TranslationTaskData {
     private final int dx;
