+++ movecraft-core/src/main/java/net/countercraft/movecraft/utils/BlockUtils.java
@@ -17,6 +17,8 @@
 
 package net.countercraft.movecraft.utils;
 
+import net.countercraft.movecraft.api.Rotation;
+
 import java.util.Arrays;
 
 public final class BlockUtils {
