+++ movecraft-core/src/main/java/net/countercraft/movecraft/utils/BoundingBoxUtils.java
@@ -17,6 +17,8 @@
 
 package net.countercraft.movecraft.utils;
 
+import net.countercraft.movecraft.api.MovecraftLocation;
+
 public final class BoundingBoxUtils {
 
     public static int[][][] formBoundingBox(MovecraftLocation[] blockList, Integer minX, Integer maxX, Integer minZ,
