+++ movecraft-core/src/main/java/net/countercraft/movecraft/utils/MapUpdateCommand.java
@@ -17,6 +17,8 @@
 
 package net.countercraft.movecraft.utils;
 
+import net.countercraft.movecraft.api.MovecraftLocation;
+import net.countercraft.movecraft.api.Rotation;
 import net.countercraft.movecraft.craft.Craft;
 
 /**
