+++ movecraft-core/src/main/java/net/countercraft/movecraft/utils/MathUtils.java
@@ -17,6 +17,8 @@
 
 package net.countercraft.movecraft.utils;
 
+import net.countercraft.movecraft.api.MovecraftLocation;
+import net.countercraft.movecraft.api.Rotation;
 import org.bukkit.Location;
 
 public final class MathUtils {
